+++ src/bittytax/conv/parsers/kraken.py
@@ -168,7 +168,7 @@ def _parse_kraken_ledgers_row(
                 dup_data_row = copy.copy(data_row)
                 dup_data_row.row = []
                 dup_data_row.t_record = TransactionOutRecord(
-                    TrType.GIFT_RECEIVED,
+                    TrType.FEE_REBATE,
                     data_row.timestamp,
                     buy_quantity=abs(Decimal(row_dict["fee"])),
                     buy_asset=_normalise_asset(row_dict["asset"]),
@@ -200,7 +200,7 @@ def _parse_kraken_ledgers_row(
                 dup_data_row = copy.copy(data_row)
                 dup_data_row.row = []
                 dup_data_row.t_record = TransactionOutRecord(
-                    TrType.GIFT_RECEIVED,
+                    TrType.FEE_REBATE,
                     data_row.timestamp,
                     buy_quantity=abs(Decimal(row_dict["fee"])),
                     buy_asset=_normalise_asset(row_dict["asset"]),
@@ -210,7 +210,7 @@ def _parse_kraken_ledgers_row(
                 data_rows.insert(row_index + 1, dup_data_row)
     elif row_dict["type"] == "invite bonus":
         data_row.t_record = TransactionOutRecord(
-            TrType.GIFT_RECEIVED,
+            TrType.REFERRAL,
             data_row.timestamp,
             buy_quantity=Decimal(row_dict["amount"]),
             buy_asset=_normalise_asset(row_dict["asset"]),